Added run_command service to simulations (incomplete)
run_command allows commands to be sent to and executed by a
simulation (or the simulation runner), without needing a
special GUI. This will make it easier to add new features in the future,
especially lightly used features that don't need GUI for use.

There needs to be more work with regards to keeping the GUI
in sync with the current state of the simulation/runner. For example,
if the simulation is started or stopped via a command, the GUI
will not reflect that.