SimulationServer is now owner- and editor-aware.
SimulationServer now requires tokens for certain commands, and
will verify that the token is for an editor or the owner
of the process.

This allows clients to connect directly to the simulation server,
while enforcing that only one of them can edit the simulation
at a time, and also allowing the timelines service to control
who is allowed to edit the simulation, and under what circumstances.